refactor[cuda]: compose pooled file read backends
